pyproject.toml
[tool.poetry]
name = "zenn-hack-backend"
version = "0.1.0"
description = "Zenn Hack Backend"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pydantic = "^1.8.2"
firebase-admin = "^5.0.0"
google-generativeai = "^0.3.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 